REFRACTOR : create package frame
