+++ src/org/analyse/core/gui/windowfocus.java
@@ -5,6 +5,8 @@
  */
 package org.analyse.core.gui;
 
+import org.analyse.core.gui.frame.AnalyseFrame;
+
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseListener;
 import java.awt.event.WindowEvent;
